backend/requirements.txt
openai==1.34.0
Flask>=2.0.0
web3==7.2.0
requests
python-dotenv
flask-cors==5.0.0